// acad-front/src/app/category/models/asistencia-model.ts
export class AsistenciaModel {
  idAsistencia?: number;
  fechaRegistroAsistencia?: string;
  usuarioIdUsuario?: number;
  usuarioNombre?: string;
  cursoIdCurso?: number;
  cursoNombre?: string;
  docenteIdDocente?: number;
  docenteNombre?: string;
  estadoAsistencia?: string;
  planAcademicoIdPlanAcademico?: number;
  planAcademicoNombre?: string;
}